Fold TeiMergerTest into PlatformMergerTest
