--- pipeline/merger/src/test/scala/weco/pipeline/merger/services/TeiMergerTest.scala
@@ -1,68 +1,0 @@
-package weco.pipeline.merger.services
-
-import org.scalatest.funspec.AnyFunSpec
-import org.scalatest.matchers.should.Matchers
-import weco.catalogue.internal_model.work.WorkState.Merged
-import weco.catalogue.internal_model.work.generators.SourceWorkGenerators
-import weco.catalogue.internal_model.work.{InternalWork, Work, WorkData}
-import weco.pipeline.matcher.generators.MergeCandidateGenerators
-
-// Tests here will eventually be folded into PlatformMergerTest.scala
-// This allows us to test using the TeiOnMerger until we make that the default merger
-class TeiMergerTest extends AnyFunSpec with SourceWorkGenerators with MergeCandidateGenerators with Matchers {
-
-  it("merges a physical sierra with a tei") {
-    val merger = PlatformMerger
-    val physicalWork =
-      sierraIdentifiedWork()
-        .items(List(createIdentifiedPhysicalItem))
-    val teiWork = teiIdentifiedWork()
-      .mergeCandidates(List(createSierraPairMergeCandidateFor(physicalWork)))
-
-    val result = merger
-      .merge(works = Seq(teiWork, physicalWork))
-      .mergedWorksWithTime(now)
-    val redirectedWorks = result.collect {
-      case w: Work.Redirected[Merged] => w
-    }
-    val visibleWorks = result.collect { case w: Work.Visible[Merged] => w }
-
-    redirectedWorks should have size 1
-    visibleWorks should have size 1
-
-    visibleWorks.head.state.canonicalId shouldBe teiWork.state.canonicalId
-  }
-
-  it("copies the thumbnail to the inner works") {
-    val teiWork = teiIdentifiedWork()
-      .mapState {
-        _.copy(
-          internalWorkStubs = List(
-            InternalWork.Identified(
-              sourceIdentifier = createTeiSourceIdentifier,
-              canonicalId = createCanonicalId,
-              workData = WorkData(
-                title = Some(s"tei-inner-${randomAlphanumeric(length = 10)}")
-              )
-            )
-          )
-        )
-      }
-
-    val sierraWork = sierraPhysicalIdentifiedWork()
-
-    val metsWork = metsIdentifiedWork()
-      .thumbnail(createDigitalLocation)
-
-    val result =
-      PlatformMerger
-        .merge(works = Seq(teiWork, sierraWork, metsWork))
-        .mergedWorksWithTime(now)
-
-    val visibleWorks = result.collect { case w: Work.Visible[Merged] => w }
-
-    visibleWorks.foreach {
-      _.data.thumbnail shouldBe metsWork.data.thumbnail
-    }
-  }
-}
